requirements.txt
langchain 
pandas
tiktoken 
huggingface_hub 
langchain-community 
langchain-milvus 
faiss-cpu
sentence-transformers 
python-docx 
odfpy 
langgraph 
"pymilvus[model]"
vllm
psycopg2-binary
aiogram 
whisperx 
requests 
pydub